Add files via upload
Women Cloth Reviews Prediction with Multi Nominal Navive Bayes.Multinomial Naïve Bayes algorithm is a probabilistic learning method that is mostly used in Natural Language Processing. It is also suitable for text classification with discrete feature. In this project it will be used to build a women cloth review prediction model.